backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1

# LLM & Agents
groq==0.9.0
langchain==0.2.5
langchain-community==0.2.5
langchain-groq==0.1.5
langgraph==0.1.14

# Embeddings & Vector
sentence-transformers==3.0.1
supabase==2.5.0
pgvector==0.3.1

# Document parsing
PyMuPDF==1.24.5
python-docx==1.1.2
youtube-transcript-api==0.6.2

# Utilities
redis==5.0.6
celery==5.4.0
tenacity==8.3.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
